# docs/api/arx_backend_api_spec.yaml
openapi: 3.0.3
info:
  title: Arx Backend API
  description: |
    Comprehensive API for the Arx Backend service providing BIM object management,
    asset inventory, CMMS integration, maintenance workflows, and compliance reporting.

    ## Authentication
    The API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting
    - Public endpoints: 100 requests/second, burst of 200
    - Authenticated endpoints: 50 requests/second, burst of 100

    ## Error Handling
    All endpoints return standard HTTP status codes with detailed error messages in JSON format.
  version: 1.0.0
  contact:
    name: Arx Development Team
    email: dev@arxos.com
  license:
    name: Proprietary
    url: https://arxos.com/license
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://api.arxos.com/api
    description: Production server
...
paths:
...
  /buildings/{id}/floors:
    get:
      tags:
        - Buildings
      summary: List building floors
      description: Retrieve all floors for a specific building
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of floors
          content:
            application/json:
              schema:
                type: object
                properties:
                  floors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Floor'

  # BIM Objects - Walls
...
  /routes:
    post:
      tags:
        - Routes
      summary: Create route
      description: Create a new route
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - start_point
                - end_point
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                start_point:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                    floor_id:
                      type: integer
                end_point:
                  type: object
                  properties:
                    x:
                      type: number
                    y:
                      type: number
                    floor_id:
                      type: integer
                route_type:
                  type: string
                  enum: [evacuation, maintenance, visitor, emergency]
                waypoints:
                  type: array
                  items:
                    type: object
                    properties:
                      x:
                        type: number
                      y:
                        type: number
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'

  # Object Info and SVG
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Floor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        building_id:
          type: integer
        svg_data:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Route:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        start_point:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            floor_id:
              type: integer
        end_point:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            floor_id:
              type: integer
        route_type:
          type: string
          enum: [evacuation, maintenance, visitor, emergency]
        waypoints:
          type: array
          items:
            type: object
            properties:
              x:
                type: number
              y:
                type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...